--- ci/.darglint
@@ -1,2 +1,0 @@
-[darglint]
-strictness=short
