+++ setup.cfg
@@ -10,11 +10,7 @@ per-file-ignores =
   src/imitation/envs/examples/airl_envs/*.py:D
 
 [darglint]
-# This section just disables darglint when run with flake8.
-# This is because darglint is slow, so we don't want to run
-# it by default. We have a separate config in ci/.darglint
-# that enables it when run from that directory. 
-ignore=*
+strictness=short
 
 [isort]
 known_first_party=imitation
