+++ src/imitation/rewards/reward_nets.py
@@ -9,7 +9,6 @@ import torch as th
 from stable_baselines3.common import preprocessing
 from torch import nn
 
-import imitation.rewards.common as rewards_common
 from imitation.util import networks
 
 
@@ -75,16 +74,36 @@ class RewardNet(nn.Module, abc.ABC):
             Preprocessed transitions: a Tuple of tensors containing
             observations, actions, next observations and dones.
         """
-        return rewards_common.disc_rew_preprocess_inputs(
-            observation_space=self.observation_space,
-            action_space=self.action_space,
-            state=state,
-            action=action,
-            next_state=next_state,
-            done=done,
-            device=self.device,
-            normalize_images=self.normalize_images,
+        state_th = th.as_tensor(state, device=self.device)
+        action_th = th.as_tensor(action, device=self.device)
+        next_state_th = th.as_tensor(next_state, device=self.device)
+        done_th = th.as_tensor(done, device=self.device)
+
+        del state, action, next_state, done  # unused
+
+        # preprocess
+        state_th = preprocessing.preprocess_obs(
+            state_th,
+            self.observation_space,
+            self.normalize_images,
         )
+        action_th = preprocessing.preprocess_obs(
+            action_th,
+            self.action_space,
+            self.normalize_images,
+        )
+        next_state_th = preprocessing.preprocess_obs(
+            next_state_th,
+            self.observation_space,
+            self.normalize_images,
+        )
+        done_th = done_th.to(th.float32)
+
+        n_gen = len(state_th)
+        assert state_th.shape == next_state_th.shape
+        assert len(action_th) == n_gen
+
+        return state_th, action_th, next_state_th, done_th
 
     def predict(
         self,
